+++ src/app/modules/resource-table/resource-table.component.ts
@@ -136,13 +136,16 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
             this.scrollViewport.scrollToIndex(0);
             this.filterChanged.next(value);
           } else {
-            this.dataSource.filter = { ...value } as string;
-            // setTimeout is needed to update the table after this.dataSource changes
-            setTimeout(() => this.onScroll());
+            this.setClientFilter({ ...value });
           }
         })
     );
   }
+
+  /**
+   * Tooltip text for a record checkbox.
+   */
+  @Input() checkboxTooltipText: string;
 
   /**
    * Get loading message according to loading status
@@ -333,7 +336,8 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
         this.liveAnnoncer.announce(
           `The ${this.resourceType} resources loading process has finished. ` +
             `${this.resources.length} rows loaded. ` +
-            this.getSortMessage()
+            this.getSortMessage(),
+          'assertive'
         );
         this.progressBarPosition$ = null;
       }
@@ -392,20 +396,28 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
 
     // Update resource table columns
     if (changes['columnDescriptions'] && this.columnDescriptions) {
+      const scrollViewport = this.scrollViewport?.elementRef.nativeElement;
+
       this.columns.length = 0;
       if (this.enableSelection || this.context === 'browse') {
         this.columns.push('select');
       }
 
-      this.setTableColumns();
+      this.setTableColumns(
+        !!scrollViewport?.querySelector(
+          '.cdk-keyboard-focused .mat-sort-header-container'
+        )
+      );
     }
   }
 
   /**
    * Set material table columns from column descriptions.
    * Set up filter form and table filtering logic if client filtering is enabled.
+   * @param isSortHeaderFocused - is there focus on the column header and should
+   *  it be maintained after redrawing the table header.
    */
-  setTableColumns(): void {
+  setTableColumns(isSortHeaderFocused: boolean): void {
     this.columns = this.columns.concat(
       this.columnDescriptions.map((c) => c.element)
     );
@@ -441,6 +453,19 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
         strategy.stickyChange.next(
           strategy.viewport.getOffsetToRenderedContentStart()
         );
+      }
+      if (isSortHeaderFocused) {
+        const scrollViewport = this.scrollViewport.elementRef.nativeElement;
+        const toFocus =
+          // Focus on the sorted column header
+          scrollViewport.querySelector<HTMLElement>(
+            `.mat-column-${this.sort.active} .mat-sort-header-container`
+          ) ||
+          // otherwise (if the sorted column is hidden) on the first column header
+          scrollViewport.querySelector<HTMLElement>(
+            `.mat-sort-header-container`
+          );
+        toFocus?.focus();
       }
     });
   }
@@ -606,7 +631,10 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
     const header = columnDescriptions
       .map((columnDescription) => columnDescription.displayName)
       .join(',');
-    const rows = this.dataSource.data.map((row) =>
+    const rowsToDownload = this.enableFiltering
+      ? this.dataSource.filteredData
+      : this.dataSource.data;
+    const rows = rowsToDownload.map((row) =>
       columnDescriptions
         .map((columnDescription) => {
           const cellText = row.cells[columnDescription.element];
@@ -777,11 +805,29 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
    * Returns a sort message.
    */
   getSortMessage(): string {
-    return this.sort?.active
-      ? `The data was sorted by ${this.sort.active} in ${
-          // MatTable shows sort order icons in reverse (see comment to PR on LF-1905).
-          this.sort.direction === 'desc' ? 'ascending' : 'descending'
-        } order.`
-      : '';
+    let message = '';
+    if (this.sort?.active) {
+      const sortingColumnDescription = this.columnDescriptionsService
+        .getAvailableColumns(this.resourceType, this.context)
+        .find((c) => c.element === this.sort.active);
+      message = `The data was sorted by ${
+        sortingColumnDescription.displayName
+      } in ${
+        // MatTable shows sort order icons in reverse (see comment to PR on LF-1905).
+        this.sort.direction === 'desc' ? 'ascending' : 'descending'
+      } order.`;
+    }
+
+    return message;
+  }
+
+  /**
+   * Sets client-side filter values
+   * @param value - filter values
+   */
+  setClientFilter(value: any): void {
+    this.dataSource.filter = { ...value } as string;
+    // setTimeout is needed to update the table after this.dataSource changes
+    setTimeout(() => this.onScroll());
   }
 }