+++ src/app/modules/select-records-page/select-records-page.component.spec.ts
@@ -25,6 +25,20 @@ describe('SelectRecordsPageComponent', () => {
   let fixture: ComponentFixture<SelectRecordsPageComponent>;
   let mockHttp: HttpTestingController;
   let loader: HarnessLoader;
+  const statuses = [
+    'candidate',
+    'eligible',
+    'follow-up',
+    'ineligible',
+    'not-registered',
+    'off-study',
+    'on-study',
+    'on-study-intervention',
+    'on-study-observation',
+    'pending-on-study',
+    'potential-candidate',
+    'screening,withdrawn'
+  ];
 
   beforeEach(async () => {
     await configureTestingModule(
@@ -35,7 +49,8 @@ describe('SelectRecordsPageComponent', () => {
       {
         features: {
           hasResearchStudy: true
-        }
+        },
+        serverUrl: 'https://dbgap-api.ncbi.nlm.nih.gov/fhir/x1'
       }
     );
     mockHttp = TestBed.inject(HttpTestingController);
@@ -61,7 +76,11 @@ describe('SelectRecordsPageComponent', () => {
     await fixture.whenStable();
     fixture.detectChanges();
     mockHttp
-      .expectOne('$fhir/ResearchStudy?_count=50&_sort=title')
+      .expectOne(
+        `$fhir/ResearchStudy?_count=50&_has:ResearchSubject:study:status=${statuses.join(
+          ','
+        )}&_sort=title`
+      )
       .flush(researchStudies);
     await expectNumberOfRecords(2);
   }
@@ -144,25 +163,25 @@ describe('SelectRecordsPageComponent', () => {
     );
     await checkBox.check();
     // No studies in the cart
-    expect(studyCart.records.length).toEqual(0);
+    expect(studyCart.listItems.length).toEqual(0);
     // Add the selected study to the cart
     await addButton.click();
     // One study in the cart
-    expect(studyCart.records.length).toEqual(1);
-    expect(studyCart.records[0].id).toEqual('phs001603.v1.p1');
+    expect(studyCart.listItems.length).toEqual(1);
+    expect(studyCart.listItems[0].id).toEqual('phs001603.v1.p1');
 
-    await loadVariables('study_id:(phs001603.v1.p1)');
+    await loadVariables('study_id:(phs001603.v1.p1*)');
 
     await selectTab('Studies');
     const removeButton = await loader.getHarness(
       MatButtonHarness.with({
-        selector: '.mat-tab-body-active app-cart .remove-btn'
+        selector: '.mat-tab-body-active app-cart .list-toolbar__icon button'
       })
     );
     // Remove the study from the cart
     await removeButton.click();
     // No studies in the cart
-    expect(studyCart.records.length).toEqual(0);
+    expect(studyCart.listItems.length).toEqual(0);
 
     await loadVariables();
   });